poprawki, jeszcze duze ujemne tablice, div, mod

+++ kompilator/src/gram.tab.h
@@ -97,7 +97,7 @@ union YYSTYPE
 {
 #line 16 "gram.y"
  
-    int intval;
+    i64 intval;
     Node* node;
     Identifier* idf;
     Condition* cond;
